pyproject.toml
[project]
name = "jdclothesagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.8",
    "dashscope>=1.24.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.3.31",
    "langchain-deepseek>=1.0.1",
    "langchain-text-splitters>=1.1.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.56.0",
]
